ai-service/intelligence-service/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
pypdf==4.3.1
sentence-transformers==3.0.1
chromadb==0.5.5
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-text-splitters==0.2.4
langchain-groq==0.1.9
langchain-google-genai==1.0.10
python-dotenv==1.0.1
python-multipart==0.0.9
torch==2.3.1
numpy<2.0.0
python-telegram-bot==21.0
edge-tts
